Refactor sankey diagram to handle segments separately
