+++ src/features/smartSearch/components/sankeyDiagram/makeSankeySegments.spec.ts
