+++ src/features/smartSearch/components/sankeyDiagram/types.ts
